tests: updated unit tests
